Move TestWriteCommand to test dir
Because of annotation processor issues, should be
fixed in https://issues.apache.org/jira/browse/IGNITE-17982
